+++ CHANGELOG.md
@@ -145,6 +145,7 @@ CHANGELOG
 ### Tools
 - `intelmqctl`:
   - fix process manager initialization if run non-interactively, as intelmqdump does it (PR#2189 by Sebastian Wagner, fixes 2188).
+  - `check`: handle `SyntaxError` in bot modules and report it without breaking execution (fixes #2177)
   - Privilege drop before logfile creation (PR#2277 by Sebastian Waldbauer, fixes 2176)
 - `intelmqsetup`: Revised installation of manager by building the static files at setup, not build time, making it behave more meaningful. Requires intelmq-manager >= 3.1.0 (PR#2198 by Sebastian Wagner, fixes #2197).